pyproject.toml
[project]
name = "StudyCore"
version = "0.1.0"
description = "StudyCore"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pyqt6>=6.10.2",
    "qrcode>=8.2",
    "sqlalchemy>=2.0.48",
    "fastapi>=0.110.0",
    "uvicorn>=0.23.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.3.0",
    "requests>=2.32.5",
]
